+++ apps/backoffice/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @io-services-cms/backoffice
 
+## 0.2.1
+
+### Patch Changes
+
+- 3a4463b: Update to mui-italia@1.0.0, minor packages update
+
 ## 0.2.0
 
 ### Minor Changes